isolation/requirements.txt
docker-py
flask
requests
